Use transactions while writing to etcd
