+++ api/aperture/common/info/v1/info.proto
@@ -44,7 +44,3 @@ message ProcessInfo {
 message HostInfo {
   string hostname = 1;
 }
-
-message ControllerInfo {
-  string id = 1;
-}
