+++ cmd/aperture-controller/controller.go
@@ -20,7 +20,6 @@ import (
 	"github.com/fluxninja/aperture/pkg/otel"
 	"github.com/fluxninja/aperture/pkg/otelcollector"
 	"github.com/fluxninja/aperture/pkg/platform"
-	"github.com/fluxninja/aperture/pkg/uuid"
 	"github.com/fluxninja/aperture/pkg/webhooks"
 	"github.com/fluxninja/aperture/pkg/webhooks/validation"
 )
@@ -30,7 +29,6 @@ func main() {
 		platform.Config{}.Module(),
 		otel.ProvideAnnotatedControllerConfig(),
 		fx.Provide(
-			uuid.ProvideDefault,
 			clockwork.NewRealClock,
 			classification.ProvideCMFileValidator,
 			flowcontrol.ProvideCMFileValidator,
