--- pkg/uuid/provide.go
@@ -1,5 +1,0 @@
-package uuid
-
-func ProvideDefault() Provider {
-	return NewDefaultProvider()
-}
