--- pkg/uuid/uuid_suite_test.go
@@ -1,13 +1,0 @@
-package uuid_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-)
-
-func TestUUID(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "UUID Suite")
-}
