+++ plugins/service/aperture-plugin-fluxninja/heartbeats/heartbeats.go
@@ -6,6 +6,7 @@ import (
 	"net/http"
 	"time"
 
+	guuid "github.com/google/uuid"
 	clientv3 "go.etcd.io/etcd/client/v3"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/metadata"
@@ -15,9 +16,7 @@ import (
 	"google.golang.org/protobuf/types/known/durationpb"
 	"google.golang.org/protobuf/types/known/emptypb"
 	"google.golang.org/protobuf/types/known/wrapperspb"
-	"gopkg.in/yaml.v2"
 
-	infov1 "github.com/fluxninja/aperture/api/gen/proto/go/aperture/common/info/v1"
 	peersv1 "github.com/fluxninja/aperture/api/gen/proto/go/aperture/common/peers/v1"
 	heartbeatv1 "github.com/fluxninja/aperture/api/gen/proto/go/aperture/plugins/fluxninja/v1"
 	"github.com/fluxninja/aperture/pkg/agentinfo"
@@ -31,7 +30,6 @@ import (
 	"github.com/fluxninja/aperture/pkg/peers"
 	"github.com/fluxninja/aperture/pkg/status"
 	"github.com/fluxninja/aperture/pkg/utils"
-	"github.com/fluxninja/aperture/pkg/uuid"
 	"github.com/fluxninja/aperture/plugins/service/aperture-plugin-fluxninja/pluginconfig"
 )
 
@@ -59,7 +57,7 @@ type heartbeats struct {
 	clientConn       *grpc.ClientConn
 	statusRegistry   *status.Registry
 	entityCache      *entitycache.EntityCache
-	controllerInfo   *infov1.ControllerInfo
+	controllerInfo   *heartbeatv1.ControllerInfo
 	heartbeatsAddr   string
 	APIKey           string
 	jobName          string
@@ -105,46 +103,29 @@ func (h *heartbeats) start(ctx context.Context, in *ConstructorIn) error {
 	return nil
 }
 
-func (h *heartbeats) setupControllerInfo(ctx context.Context, etcdClient *etcdclient.Client, UUIDProvider uuid.Provider) error {
+func (h *heartbeats) setupControllerInfo(ctx context.Context, etcdClient *etcdclient.Client) error {
 	etcdPath := "/fluxninja/controllerid"
+	controllerID := guuid.NewString()
 
-	// check if controller id is already present
-	getResp, err := etcdClient.KV.Get(clientv3.WithRequireLeader(ctx), etcdPath, clientv3.WithPrefix(), clientv3.WithLease(etcdClient.LeaseID))
+	txn := etcdClient.Client.Txn(etcdClient.Client.Ctx())
+	resp, err := txn.If(clientv3.Compare(clientv3.CreateRevision(etcdPath), "=", 0)).
+		Then(clientv3.OpPut(etcdPath, controllerID)).
+		Else(clientv3.OpGet(etcdPath)).Commit()
 	if err != nil {
-		log.Error().Err(err).Str("etcdPath", etcdPath).Msg("Failed to list controller id")
+		log.Error().Err(err).Msg("Could not read/write controller id to etcd")
 		return err
 	}
 
-	if getResp.Count == 1 {
-		for _, kv := range getResp.Kvs {
-			controllerInfo := &infov1.ControllerInfo{}
-			err = config.Unmarshal(kv.Value, controllerInfo)
-			if err != nil {
-				log.Warn().Err(err).Msg("Failed to unmarshal controller info")
-				return err
-			}
-
-			h.controllerInfo = controllerInfo
-			return nil
+	// Succeeded is true if the If condition above is true - meaning there were no controller Id in etcd
+	if !resp.Succeeded {
+		for _, res := range resp.Responses {
+			controllerID = string(res.GetResponseRange().Kvs[0].Value)
+			break
 		}
 	}
 
-	// not present, create it
-	controllerID := UUIDProvider.New()
-	h.controllerInfo = &infov1.ControllerInfo{
+	h.controllerInfo = &heartbeatv1.ControllerInfo{
 		Id: controllerID,
-	}
-
-	dat, err := yaml.Marshal(h.controllerInfo)
-	if err != nil {
-		log.Error().Err(err).Msg("Failed to marshal controller info")
-		return err
-	}
-
-	_, err = etcdClient.KV.Put(clientv3.WithRequireLeader(ctx), etcdPath, string(dat), clientv3.WithLease(etcdClient.LeaseID))
-	if err != nil {
-		log.Error().Err(err).Msg("Failed to save controller info to etcd")
-		return err
 	}
 
 	return nil
